+++ architect_py/async_client.py
@@ -12,13 +12,12 @@ After sending the order, this string can be used to retrieve the order status
 send_limit_order -> get_order
 
 The individual graphql types are subject to change, so it is not recommended to use them directly.
-
 """
 
 import logging
 from datetime import date, datetime
 from decimal import Decimal
-from typing import Any, AsyncIterator, List, Optional, Sequence
+from typing import Any, AsyncIterator, List, Optional, Sequence, cast
 
 from architect_py.graphql_client.get_fills_query import (
     GetFillsQueryFolioHistoricalFills,
@@ -45,6 +44,7 @@ from .graphql_client.fragments import (
     CandleFields,
     ExecutionInfoFields,
     L2BookFields,
+    MarketStatusFields,
     MarketTickerFields,
     OrderFields,
     ProductInfoFields,
@@ -78,10 +78,30 @@ class AsyncClient:
         api_secret: str,
         host: str = "https://app.architect.co",
         paper_trading: bool = True,
-        port: Optional[int] = None,
+        _port: Optional[int] = None,
         **kwargs: Any,
     ):
-        """ """
+        """
+        Args:
+            api_key: API key for the user
+            api_secret: API secret for the user
+            host: Host for the GraphQL server, defaults to "https://app.architect.co"
+            paper_trading: Whether to use the paper trading environment, defaults to True
+            _port: Port for the GraphQL server, more for debugging purposes, do not set this unless you are sure of the port
+
+        the API key and secret can be generated on the app.architect.co website
+
+        Returns:
+            Client object
+
+        Raises:
+            ValueError: If the API key or secret are not the correct length or contain invalid characters
+
+        For any request, if you get a "GraphQLClientHttpError: HTTP status code: 500" it likely means that your
+        API key and secret are incorrect. Please double check your credentials.
+
+        If you get a "GraphQLClientHttpError: HTTP status code: 400", please contact support so we can fix the function.
+        """
 
         if not api_key.isalnum():
             raise ValueError(
@@ -100,22 +120,22 @@ class AsyncClient:
                 "API key and secret are not the correct length, please double check your credentials."
             )
 
-        if port is None:
+        if _port is None:
             if paper_trading:
-                port = 5678
+                _port = 5678
             else:
-                port = 4567
+                _port = 4567
 
         self.graphql_client = GraphQLClient(
-            api_key=api_key, api_secret=api_secret, host=host, port=port, **kwargs
+            api_key=api_key, api_secret=api_secret, host=host, port=_port, **kwargs
         )
 
         self.grpc_client = GRPCClient(self.graphql_client)
 
     async def who_am_i(self) -> tuple[str, str]:
         """
-        Returns
-        (userId, user_email)
+        Returns:
+            (user_id, user_email)
         """
         user_id = await self.graphql_client.user_id_query()
         email = await self.graphql_client.user_email_query()
@@ -128,7 +148,13 @@ class AsyncClient:
         execution_venue: Optional[str] = None,
     ) -> List[TradableProduct]:
         """
-        returns a list of TradableProduct objects
+        Args:
+            search_string: a string to search for in the symbol
+                Examples: "ES", "NQ", "GC"
+            execution_venue: the execution venue to search in
+                Examples: "CME"
+        Returns:
+            a list of TradableProduct objects
         """
         markets = (
             await self.graphql_client.search_symbols_query(
@@ -139,42 +165,112 @@ class AsyncClient:
         return markets
 
     async def get_product_info(self, symbol: str) -> Optional[ProductInfoFields]:
-        # this reduces the indirection count
+        """
+        Args:
+            symbol: the symbol to get information for
+        Returns:
+            ProductInfoFields object if the symbol exists
+        """
         info = await self.graphql_client.get_product_info_query(symbol)
         return info.product_info
 
     async def get_product_infos(
         self, symbols: list[str]
     ) -> Sequence[ProductInfoFields]:
+        """
+        Args:
+            symbols: the symbols to get information for
+        Returns:
+            a list of ProductInfoFields
+
+        Any duplicate symbols will be ignored.
+        Any invalid symbols will be ignored.
+        The order of the symbols in the list will not necessarily be preserved in the output.
+        """
         infos = await self.graphql_client.get_product_infos_query(symbols)
         return infos.product_infos
 
+    async def get_execution_info(
+        self, symbol: TradableProduct, execution_venue: str
+    ) -> Optional[ExecutionInfoFields]:
+        """
+        Args:
+            symbol: the symbol to get execution information for
+            execution_venue: the execution venue e.g. "CME"
+
+        Returns:
+            ExecutionInfoFields
+
+        Raises
+        """
+        execution_info = await self.graphql_client.get_execution_info_query(
+            symbol, execution_venue
+        )
+        return execution_info.execution_info
+
+    async def get_execution_infos(
+        self, symbols: list[TradableProduct], execution_venue: Optional[str] = None
+    ) -> Sequence[ExecutionInfoFields]:
+        """
+        Args:
+            symbols: the symbols to get execution information for
+            execution_venue: the execution venue e.g. "CME"
+
+        Returns:
+            a list of ExecutionInfoFields
+        """
+        execution_infos = await self.graphql_client.get_execution_infos_query(
+            symbols, execution_venue
+        )
+        return execution_infos.execution_infos
+
     async def get_cme_first_notice_date(self, symbol: str) -> Optional[date]:
+        """
+        Args:
+            symbol: the symbol to get the first notice date for a CME future
+
+        Returns:
+            the first notice date as a date object if it exists
+        """
         notice = await self.graphql_client.get_first_notice_date_query(symbol)
         if notice is None or notice.product_info is None:
             return None
         return notice.product_info.first_notice_date
 
     async def get_future_series(self, series_symbol: str) -> list[str]:
+        """
+        Args:
+            series_symbol: the symbol to get the series for
+                e.g. ES CME Futures" would yield a list of all the ES futures
+        Returns:
+            a list of symbols in the series
+        """
         futures_series = await self.graphql_client.get_future_series_query(
             series_symbol
         )
         return futures_series.futures_series
 
-    async def get_execution_info(
-        self, symbol: TradableProduct, execution_venue: str
-    ) -> ExecutionInfoFields:
-        execution_info = await self.graphql_client.get_execution_info_query(
-            symbol, execution_venue
-        )
-        return execution_info.execution_info
-
     @staticmethod
     def get_expiration_from_CME_name(name: str) -> date:
+        """
+        Args:
+            name: the CME future name
+                e.g. "ES 20211217 CME Future" -> date(2021, 12, 17)
+        Returns:
+            the expiration date as a date object
+        """
         _, d, *_ = name.split(" ")
         return datetime.strptime(d, "%Y%m%d").date()
 
     async def get_cme_futures_series(self, series: str) -> list[tuple[date, str]]:
+        """
+        Args:
+            series: the series to get the futures for
+                e.g. "ES CME Futures"
+        Returns:
+            a list of tuples of the expiration date and
+            the symbol for each future in the series
+        """
         markets = await self.get_future_series(
             series,
         )
@@ -190,20 +286,44 @@ class AsyncClient:
     async def get_cme_future_from_root_month_year(
         self, root: str, month: int, year: int
     ) -> str:
+        """
+        Args:
+            root: the root symbol for the future e.g. "ES"
+            month: the month of the future
+            year: the year of the future
+        Returns:
+            the symbol for the future
+
+            Errors if the result is not unique
+            This is a simple wrapper around search_symbols
+        """
         [market] = [
             market
             for market in await self.search_symbols(
                 f"{root} {year}{month:02d}",
+                execution_venue="CME",
             )
         ]
 
         return market
 
     async def list_accounts(self) -> Sequence[AccountWithPermissionsFields]:
+        """
+        Returns:
+            a list of AccountWithPermissionsFields for the user that the API key belongs to
+            (use who_am_i to get the user_id / email)
+        """
         accounts = await self.graphql_client.list_accounts_query()
         return accounts.accounts
 
     async def get_account_summary(self, account: str) -> AccountSummaryFields:
+        """
+        Args:
+            account: the account to get the summary for,
+                can be the account id( (a UUID) or the account name (e.g. CQG:00000)
+        Returns:
+            AccountSummaryFields for the account
+        """
         summary = await self.graphql_client.get_account_summary_query(account=account)
         return summary.account_summary
 
@@ -212,6 +332,16 @@ class AsyncClient:
         accounts: Optional[list[str]] = None,
         trader: Optional[str] = None,
     ) -> Sequence[AccountSummaryFields]:
+        """
+        Args:
+            accounts: a list of account ids to get summaries for,
+                can be the account id( (a UUID) or the account name (e.g. CQG:00000)
+            trader: the trader / userId to get summaries for
+
+            if both arguments are given, the accounts are all appended and returned together
+        Returns:
+            a list of AccountSummary for the accounts
+        """
         summaries = await self.graphql_client.get_account_summaries_query(
             trader=trader, accounts=accounts
         )
@@ -223,6 +353,11 @@ class AsyncClient:
         from_inclusive: Optional[datetime] = None,
         to_exclusive: Optional[datetime] = None,
     ) -> Sequence[AccountSummaryFields]:
+        """
+        Returns:
+            a list of AccountSummaryFields for the account for the given dates
+            use timestamp to get the time of the of the summary
+        """
         history = await self.graphql_client.get_account_history_query(
             account=account, from_inclusive=from_inclusive, to_exclusive=to_exclusive
         )
@@ -237,6 +372,20 @@ class AsyncClient:
         symbol: Optional[str] = None,
         parent_order_id: Optional[str] = None,
     ) -> Sequence[OrderFields]:
+        """
+        Args:
+            order_ids: a list of order ids to get
+            venue: the venue to get orders for
+            account: the account to get orders for
+            trader: the trader to get orders for
+            symbol: the symbol to get orders for
+            parent_order_id: the parent order id to get orders for
+
+            these filters are combinewd via OR statements so if you pass
+            in multiple arguments, it will return the union of the results
+        Returns:
+            a list of OrderFields of the open orders that match the union of the filters
+        """
         orders = await self.graphql_client.get_open_orders_query(
             venue=venue,
             account=account,
@@ -248,6 +397,10 @@ class AsyncClient:
         return orders.open_orders
 
     async def get_all_open_orders(self) -> Sequence[OrderFields]:
+        """
+        Returns:
+            a list of OrderFields of all the open orders for the user
+        """
         orders = await self.graphql_client.get_open_orders_query()
         return orders.open_orders
 
@@ -260,6 +413,22 @@ class AsyncClient:
         account: Optional[str] = None,
         parent_order_id: Optional[str] = None,
     ) -> Sequence[OrderFields]:
+        """
+        Args:
+            order_ids: a list of order ids to get
+            from_inclusive: the start date to get orders for
+            to_exclusive: the end date to get orders for
+            venue: the venue to get orders for, e.g. CME
+            account: the account to get orders for,
+                can be the account id( (a UUID) or the account name (e.g. CQG:00000)
+            parent_order_id: the parent order id to get orders for
+        Returns:
+            a list of OrderFields of the historical orders that match the filters
+
+            either the order_ids parameter needs to be filled
+            OR
+            the from_inclusive and to_exclusive parameters need to be filled
+        """
         orders = await self.graphql_client.get_historical_orders_query(
             order_ids=order_ids,
             venue=venue,
@@ -271,6 +440,17 @@ class AsyncClient:
         return orders.historical_orders
 
     async def get_order(self, order_id: str) -> Optional[OrderFields]:
+        """
+        Returns the OrderFields object for the specified order
+        Useful for looking at past sent orders
+
+        Args:
+            order_id: the order id to get
+        Returns:
+            the OrderFields object for the order
+
+        Queries open_orders first then queries historical_orders
+        """
         open_orders = await self.graphql_client.get_open_orders_query(
             order_ids=[order_id]
         )
@@ -287,6 +467,16 @@ class AsyncClient:
             return historical_orders.historical_orders[0]
 
     async def get_orders(self, order_ids: list[str]) -> list[Optional[OrderFields]]:
+        """
+        Returns a list of OrderFields objects for the specified orders
+        Useful for looking at past sent orders
+        Args:
+            order_ids: a list of order ids to get
+        Returns:
+            a list of OrderFields objects for the orders
+
+        Plural form of get_order
+        """
         orders_dict: dict[str, Optional[OrderFields]] = {
             order_id: None for order_id in order_ids
         }
@@ -319,24 +509,65 @@ class AsyncClient:
         account: Optional[str] = None,
         order_id: Optional[str] = None,
     ) -> GetFillsQueryFolioHistoricalFills:
+        """
+        returns a list of fills for the given filters
+        Args:
+            from_inclusive: the start date to get fills for
+            to_exclusive: the end date to get fills for
+            venue: the venue to get fills for, e.g. "CME"
+            account: the account to get fills for,
+                can be the account id( (a UUID) or the account name (e.g. CQG:00000)
+            order_id: the order id to get fills for
+        Returns:
+            a list of GetFillsQueryFolioHistoricalFills
+        """
         fills = await self.graphql_client.get_fills_query(
             venue, account, order_id, from_inclusive, to_exclusive
         )
         return fills.historical_fills
 
-    async def get_market_status(self, symbol: str, venue: str):
+    async def get_market_status(
+        self, symbol: TradableProduct, venue: str
+    ) -> MarketStatusFields:
+        """
+        Returns market status for symbol (ie if it is quoting and trading)
+        Args:
+            symbol: the symbol to get the market status for, e.g. "ES 20250321 CME Future/USD"
+            venue: the venue that the symbol is traded at, e.g. CME
+        Returns:
+            MarketStatusFields for the symbol
+        """
         market_status = await self.graphql_client.get_market_status_query(symbol, venue)
         return market_status.market_status
 
-    async def get_market_snapshot(self, symbol: str, venue: str) -> MarketTickerFields:
-        """this is an alias for l1_book_snapshot"""
+    async def get_market_snapshot(
+        self, symbol: TradableProduct, venue: str
+    ) -> MarketTickerFields:
+        """
+        this is an alias for l1_book_snapshot
+        Args:
+            symbol: the symbol to get the market snapshot for, e.g. "ES 20250321 CME Future/USD"
+            venue: the venue that the symbol is traded at, e.g. CME
+        Returns:
+            MarketTickerFields for the symbol
+        """
         return await self.get_l1_book_snapshot(symbol=symbol, venue=venue)
 
     async def get_market_snapshots(
-        self, symbols: list[str], venue: str
+        self, symbols: list[TradableProduct], venue: str
     ) -> Sequence[MarketTickerFields]:
-        """this is an alias for l1_book_snapshots"""
-        return await self.get_l1_book_snapshots(venue=venue, symbols=symbols)
+        """
+        this is an alias for l1_book_snapshots
+
+        Args:
+            symbols: the symbols to get the market snapshots for
+            venue: the venue that the symbols are traded at
+        Returns:
+            a list of MarketTickerFields for the symbols
+        """
+        return await self.get_l1_book_snapshots(
+            venue=venue, symbols=symbols  # type: ignore
+        )
 
     async def get_candles_snapshot(
         self,
@@ -346,7 +577,16 @@ class AsyncClient:
         start: datetime,
         end: Optional[datetime] = None,
     ) -> Sequence[CandleFields]:
-        start.tzinfo
+        """
+        Args:
+            symbol: the symbol to get the candles for
+            venue: the venue of the symbol
+            candle_width: the width of the candles
+            start: the start date to get candles for
+            end: the end date to get candles for
+        Returns:
+            a list of CandleFields for the specified candles
+        """
         if end is None:
             end = datetime.now(tz=start.tzinfo)
         candles = await self.graphql_client.get_candle_snapshot_query(
@@ -359,6 +599,13 @@ class AsyncClient:
         symbol: str,
         venue: str,
     ) -> MarketTickerFields:
+        """
+        Args:
+            symbol: the symbol to get the l1 book snapshot for
+            venue: the venue that the symbol is traded at
+        Returns:
+            MarketTickerFields for the symbol
+        """
         snapshot = await self.graphql_client.get_l_1_book_snapshot_query(
             symbol=symbol, venue=venue
         )
@@ -367,6 +614,13 @@ class AsyncClient:
     async def get_l1_book_snapshots(
         self, symbols: list[str], venue: str
     ) -> Sequence[MarketTickerFields]:
+        """
+        Args:
+            symbols: the symbols to get the l1 book snapshots for
+            venue: the venue that the symbols are traded at
+        Returns:
+            a list of MarketTickerFields for the symbols
+        """
         snapshot = await self.graphql_client.get_l_1_book_snapshots_query(
             venue=venue, symbols=symbols
         )
@@ -398,10 +652,10 @@ class AsyncClient:
         self,
         *,
         symbol: str,
+        execution_venue: Optional[str],
         odir: OrderDir,
         quantity: Decimal,
         limit_price: Decimal,
-        execution_venue: Optional[str],
         order_type: OrderType = OrderType.LIMIT,
         time_in_force: TimeInForce = TimeInForce.DAY,
         good_til_date: Optional[datetime] = None,
@@ -412,11 +666,32 @@ class AsyncClient:
         trigger_price: Optional[Decimal] = None,
     ) -> OrderFields:
         """
-        if execution_venue is set to None, the OMS will send the order to the primary_exchange
+        Sends a regular limit order.
 
-        the primary_exchange can be deduced from `get_product_info`
+        Args:
+            symbol: the symbol to send the order for
+            execution_venue: the execution venue to send the order to,
+                if execution_venue is set to None, the OMS will send the order to the primary_exchange
+                the primary_exchange can be deduced from `get_product_info`
+            odir: the direction of the order
+            quantity: the quantity of the order
+            limit_price: the limit price of the order
+                It is highly recommended to make this a Decimal object from the decimal module to avoid floating point errors
+            order_type: the type of the order
+            time_in_force: the time in force of the order
+            good_til_date: the date the order is good until, only relevant for time_in_force = "GTD"
+            price_round_method: the method to round the price to the nearest tick, will not round if None
+            account: the account to send the order for
+                While technically optional, for most order types, the account is required
+            trader: the trader to send the order for, defaults to the user's trader
+                for when sending order for another user, not relevant for vast majority of users
+            post_only: whether the order should be post only, not supported by all exchanges
+            trigger_price: the trigger price for the order, only relevant for stop / take_profit orders
+        Returns:
+            the OrderFields object for the order
+            The order.status should  be "PENDING" until the order is "OPEN" / "REJECTED" / "OUT" / "CANCELED" / "STALE"
 
-        While technically optional, for most order types, the account is required
+            If the order is rejected, the order.reject_reason and order.reject_message will be set
         """
         assert quantity > 0, "quantity must be positive"
 
@@ -435,6 +710,10 @@ class AsyncClient:
             execution_info = await self.get_execution_info(
                 TradableProduct(symbol), execution_venue
             )
+            if execution_info is None:
+                raise ValueError(
+                    "Could not find execution information for {symbol} for rounding price for limit order. Please round price manually."
+                )
             if (tick_size := execution_info.tick_size) is not None:
                 if tick_size:
                     limit_price = nearest_tick(
@@ -475,6 +754,25 @@ class AsyncClient:
         account: Optional[str] = None,
         fraction_through_market: Decimal = Decimal("0.001"),
     ) -> OrderFields:
+        """
+        Sends a market order with a limit price based on the BBO
+        Meant to behave as a market order but with more protections
+
+        Args:
+            symbol: the symbol to send the order for
+            execution_venue: the execution venue to send the order to
+            odir: the direction of the order
+            quantity: the quantity of the order
+            time_in_force: the time in force of the order
+            account: the account to send the order for
+            fraction_through_market: the fraction through the market to send the order at
+                e.g. 0.001 would send the order 0.1% through the market
+        Returns:
+            the OrderFields object for the order
+            The order.status should  be "PENDING" until the order is "OPEN" / "REJECTED" / "OUT" / "CANCELED" / "STALE"
+
+            If the order is rejected, the order.reject_reason and order.reject_message will be set
+        """
 
         # Check for GQL failures
         bbo_snapshot = await self.get_market_snapshot(
@@ -539,9 +837,22 @@ class AsyncClient:
         )
 
     async def cancel_order(self, order_id: str) -> CancelFields:
+        """
+        Cancels an order by order id
+        Args:
+            order_id: the order id to cancel
+        Returns:
+            the CancelFields object
+        """
         cancel = await self.graphql_client.cancel_order_mutation(order_id)
         return cancel.cancel_order
 
     async def cancel_all_orders(self) -> bool:
+        """
+        Cancels all open orders
+        Returns:
+            True if all orders were cancelled successfully
+            False if there was an error
+        """
         b = await self.graphql_client.cancel_all_orders_mutation()
         return b.cancel_all_orders